requirements.txt
openai
chromadb
sqlparse
pandas
mysql-connector-python
python-dotenv
streamlit